+++ cartridge/health.lua
@@ -1,14 +1,9 @@
 local membership = require('membership')
 local argparse = require('cartridge.argparse')
 
--- Original private member_is_healthy function:
--- https://github.com/tarantool/cartridge/blob/master/cartridge/rpc.lua
-local function is_healthy(_)
-    local member = membership.myself()
-    local parse = argparse.parse()
-    local instance = parse.instance_name or parse.alias or 'instance'
-    if box.info.status and box.info.status == 'running'
-        and member ~= nil
+local function member_is_healthy(member)
+    return (
+        (member ~= nil)
         and (member.status == 'alive' or member.status == 'suspect')
         and (
             member.payload.state_prev == nil or -- for backward compatibility with old versions
@@ -19,11 +14,33 @@ local function is_healthy(_)
             member.payload.state == 'ConfiguringRoles' or
             member.payload.state == 'RolesConfigured'
         )
-    then
+    )
+end
+
+-- TODO
+-- During the transfer of the metrics module to Tarantool, the same implementation of this function will occur
+-- several times:
+-- * https://github.com/tarantool/metrics/blob/master/cartridge/health.lua
+-- * https://github.com/tarantool/cartridge/blob/master/cartridge/health.lua
+-- * https://github.com/tarantool/cartridge/blob/master/cartridge/rpc.lua
+local function is_healthy()
+    local member = membership.myself()
+    return box.info.status and box.info.status == 'running'
+        and member_is_healthy(member)
+end
+
+local function default_is_healthy_handler(_)
+    local parse = argparse.parse()
+    local instance = parse.instance_name or parse.alias or 'instance'
+    if is_healthy() then
         return {body = instance .. " is OK", status = 200}
     else
         return {body = instance .. " is dead", status = 500}
     end
 end
 
-return {is_healthy = is_healthy}
+return {
+    member_is_healthy = member_is_healthy,
+    is_healthy = is_healthy,
+    default_is_healthy_handler = default_is_healthy_handler,
+}